+++ AAEmu.Game/Core/Managers/World/StreamManager.cs
@@ -9,11 +9,11 @@ namespace AAEmu.Game.Core.Managers.World;
 
 public class StreamManager : Singleton<StreamManager>
 {
-    private readonly Dictionary<uint, uint> _accounts;
+    private readonly Dictionary<uint, ulong> _accounts;
 
     protected StreamManager()
     {
-        _accounts = new Dictionary<uint, uint>();
+        _accounts = new Dictionary<uint, ulong>();
     }
 
     public static void Load()
@@ -21,7 +21,7 @@ public class StreamManager : Singleton<StreamManager>
         // TODO ...
     }
 
-    public void AddToken(uint accountId, uint connectionId)
+    public void AddToken(ulong accountId, uint connectionId)
     {
         _accounts.Add(connectionId, accountId);
     }
@@ -31,7 +31,7 @@ public class StreamManager : Singleton<StreamManager>
         _accounts.Remove(token);
     }
 
-    public void Login(StreamConnection connection, uint accountId, uint token)
+    public void Login(StreamConnection connection, ulong accountId, uint token)
     {
         if (_accounts.ContainsKey(token))
         {
