+++ AAEmu.Game/Core/Network/Connections/GameConnectionTable.cs
@@ -7,11 +7,11 @@ namespace AAEmu.Game.Core.Network.Connections;
 
 public class GameConnectionTable : Singleton<GameConnectionTable>
 {
-    private ConcurrentDictionary<uint, GameConnection> _connections;
+    private ConcurrentDictionary<ulong, GameConnection> _connections;
 
     private GameConnectionTable()
     {
-        _connections = new ConcurrentDictionary<uint, GameConnection>();
+        _connections = new ConcurrentDictionary<ulong, GameConnection>();
     }
 
     public void AddConnection(GameConnection con)
@@ -19,13 +19,13 @@ public class GameConnectionTable : Singleton<GameConnectionTable>
         _connections.TryAdd(con.Id, con);
     }
 
-    public GameConnection GetConnection(uint id)
+    public GameConnection GetConnection(ulong id)
     {
         _connections.TryGetValue(id, out var con);
         return con;
     }
 
-    public GameConnection RemoveConnection(uint id)
+    public GameConnection RemoveConnection(ulong id)
     {
         _connections.TryRemove(id, out var con);
         return con;
@@ -36,7 +36,7 @@ public class GameConnectionTable : Singleton<GameConnectionTable>
         return new List<GameConnection>(_connections.Values);
     }
 
-    public GameConnection GetConnectionByAccount(uint accountId)
+    public GameConnection GetConnectionByAccount(ulong accountId)
     {
         var connectionInfo = _connections.Where(c => c.Value.AccountId == accountId).ToList();
         if (connectionInfo.Count >= 1)
