+++ AAEmu.Game/Models/Game/Skills/Effects/BuffEffect.cs
@@ -46,7 +46,7 @@ public class BuffEffect : EffectTemplate
         if (target.Buffs.CheckBuffImmune(Buff.Id))
             return; // TODO send error of immune?
 
-        uint abLevel = 1;
+        ushort abLevel = 1;
         if (caster is Character character)
         {
             Logger.Debug("BuffEffect");
@@ -55,12 +55,12 @@ public class BuffEffect : EffectTemplate
                 var template = source.Skill.Template;
                 var abilityLevel = character.GetAbLevel((AbilityType)source.Skill.Template.AbilityId);
                 if (template.LevelStep != 0)
-                    abLevel = (uint)((abilityLevel / template.LevelStep) * template.LevelStep);
+                    abLevel = (ushort)((abilityLevel / template.LevelStep) * template.LevelStep);
                 else
-                    abLevel = (uint)template.AbilityLevel;
+                    abLevel = (ushort)template.AbilityLevel;
 
                 //Dont allow lower than minimum ablevel for skill or infinite debuffs can happen
-                abLevel = (uint)Math.Max(template.AbilityLevel, (int)abLevel);
+                abLevel = (ushort)Math.Max(template.AbilityLevel, abLevel);
             }
             else if (source.Buff != null)
             {
@@ -71,7 +71,7 @@ public class BuffEffect : EffectTemplate
         {
             if (source.Skill != null)
             {
-                abLevel = (uint)source.Skill.Template.AbilityLevel;
+                abLevel = (ushort)source.Skill.Template.AbilityLevel;
             }
         }
